Fix renames and add docs
