+++ packages/flame/lib/src/components/core/component_render_context.dart
